requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
pydantic[email]
passlib[bcrypt]
pyJWT
python-multipart
celery
redis
torch
torchvision
uvicorn[standard]
requests
grpcio
grpcio-tools